Part and .cfg edits
Fixed Z-fighting on the long ladder, made the little step on the small
ladder transparent.

Changed the mass and EC generation of the 1x3 solar panels to match
their size. Moved the 1x3 wide further down the tech tree.

Moved the MK16 parachute up to blend into the top of the MK1 pod.

Reduced the CoM offset on the command pods by quite a bit.