The Update to Break Them All
Restructured all of the configuration files to be more compact and located in the approprate places. Added in several hidden parts that were previously floating about in the mod. Half of the gauges won't load, but most of the parts seem to be in order. Updated MM. Removed Defunct mod compatibility files. Moved the fixed antenna from the Parts Bin folder to the Stock folder. Removed the pruner files. Probably pissed off everyone who was fixing the mod previously. May the Kraken have mercy on my soul...